+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+## [0.1.0-alpha.15] - 2026-08-27
+
 ### Added
 
 - Production reality's "Fetch from observation source" now offers a "Save this observation" action